Add more overflow fixes (sdla) Prevent unloading of old IDE drivers Fix compilation with kmod off Fix typo in NUMA API Fix LFS in sendfile (#39156) Fix bug in i810 overflow fix
suse-commit: ce12ebd175645d2687bc107bb3b0476b7b101115

+++ drivers/char/drm/i810_dma.c
@@ -1273,7 +1273,7 @@ int i810_dma_mc(struct inode *inode, struct file *filp,
 		return -EINVAL;
 	}
 
-	if (mc.idx >= dma->buf_count)
+	if ((unsigned long)mc.idx >= dma->buf_count)
 		return -EINVAL;
 
 	i810_dma_dispatch_mc(dev, dma->buflist[mc.idx], mc.used,
